move catsStateTEffect instance into stateT object
