+++ api/src/test/scala/quasar/api/datasource/MockDatasourcesSpec.scala
@@ -20,7 +20,7 @@ import slamdata.Predef.{Int, List, String, Stream => SStream}
 import quasar.Condition
 import quasar.api.MockSchemaConfig
 import quasar.contrib.cats.stateT._
-import quasar.contrib.cats.effect.catsStateTEffect
+import quasar.contrib.cats.effect.stateT.catsStateTEffect
 
 import scala.concurrent.ExecutionContext.Implicits.global
